Begin adding wrappers
The initial ObjectHandle and ObjectIterator structs will allow the user
to explore and interact with various SystemVerilog objects throughout the
simulation hierarchy

The vlog_startup_routines!() macro allows the user to specify a list of
startup functions/closures to invoke when the shared library is loaded
by a simulator.

The sv_print* macros are analogous to the standard ones but print to the
simulator's output instead.

Initial support for querying information about the simulator in use has
also been implemented.